configure: add and enable maintainer mode
This allows maintainers and distributors to disable maintainer mode and
a bunch of rules forcing a reconfigure if any of the input files (mostly
configure.ac, Makefile.am) has changed. Avoids needing autotools to be
used or even installed if the generated build environment needs to be
patched, e.g. for certain cross compilation scenarios.

Signed-off-by: Bert van Hall <bert.vanhall@gmx.de>