use a clearer error message clearer when your password is too long
Consider the password '12345678901234567890123456789012' which is 32 bytes.

Prior to this commit, you could use `x11vnc -storepasswd file` and you would
see a prompt to enter that password. When you type it in and press enter,
the prompt would turn into:
```
Enter VNC password:
Verify password:
** passwords differ.
```

The reason is related to the length of the buffer and `fgets()` is kinda
broken. It doesn't tell you that there's too many characters. But it does
leave the newline from pressing enter on input. So that gets checked for,
and if it's not there then it's assumed the password was too long.

The new output is:
```
Enter VNC password:
Verify password:
** password exceeds maximum 30 characters.
```

This makes it clearer to the user how to solve the problem.