Support shm key and drm for rawfb option
This will support both x11 and Wayland. (#227)

* Add support shm key for rawfb option

Sometime shmid is not same on different time or system, we should use a constant ID for shm. Here use shm key to instead.

* Add DRM support

access GPU with DRM, this will support x11 and wayland

* Fix a bug for shmget params.

shmget should be 0 otherwise it would alloc a new shared memory but not for the frame buffer.

* fix drm configure while compiling.